# drfblog/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
  /api/:
    get:
      operationId: listPosts
      description: ''
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
          description: ''
      tags:
      - api
    post:
      operationId: createPost
      description: ''
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Post'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
          description: ''
      tags:
      - api
...
components:
  schemas:
    Post:
      type: object
      properties:
        category:
          type: integer
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 250
        slug:
          type: string
          maxLength: 250
          pattern: ^[-a-zA-Z0-9_]+$
        author:
          type: integer
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        status:
          enum:
          - draft
          - published
          type: string
      required:
      - title
      - slug
      - author
      - content
...